requirements.txt
pandas
matplotlib
seaborn==0.13.2
scikit-learn==1.5.2
pyyaml==6.0.2
isort==5.13.2
black==24.10.0
Flask==3.1.0